requirements.txt
spacy==3.6.1
transformers==4.31.0
torch==2.0.1
scikit-learn==1.3.0
seqeval==1.2.2
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
pyyaml==6.0.1
